Ajuste menu master page

+++ WebApp_NaturalesBuenavida/Presentation/Dashboard.aspx.cs
@@ -15,7 +15,7 @@ namespace Presentation
         protected void Page_Load(object sender, EventArgs e)
         {
             Usuario usuario = Session["Usuario"] as Usuario;
-            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(Privilegios.Dashboard.ToString()))
+            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(((int)Privilegios.Dashboard).ToString()))
             {
                 Response.Redirect("AccessDenied.aspx");
             }
