+++ WebApp_NaturalesBuenavida/Presentation/WFCities.aspx.cs
@@ -20,7 +20,7 @@ namespace Presentation
             }
 
             Usuario usuario = Session["Usuario"] as Usuario;
-            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(Privilegios.Ciudades.ToString()))
+            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(((int)Privilegios.Ciudades).ToString()))
             {
                 Response.Redirect("AccessDenied.aspx");
             }
