+++ WebApp_NaturalesBuenavida/Presentation/WFPresentation.aspx.cs
@@ -25,7 +25,7 @@ namespace Presentation
                
             }
             Usuario usuario = Session["Usuario"] as Usuario;
-            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(Privilegios.Presentacion.ToString()))
+            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(((int)Privilegios.Presentacion).ToString()))
             {
                 Response.Redirect("AccessDenied.aspx");
             }
