+++ WebApp_NaturalesBuenavida/Presentation/WFPrivileges.aspx.cs
@@ -27,7 +27,7 @@ namespace Presentation
 
             }
             Usuario usuario = Session["Usuario"] as Usuario;
-            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(Privilegios.Privilegios.ToString()))
+            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(((int)Privilegios.Privilegios).ToString()))
             {
                 Response.Redirect("AccessDenied.aspx");
             }
