+++ WebApp_NaturalesBuenavida/Presentation/WFStates.aspx.cs
@@ -21,7 +21,7 @@ namespace Presentation
             }
 
             Usuario usuario = Session["Usuario"] as Usuario;
-            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(Privilegios.Departamentos.ToString()))
+            if (usuario == null || usuario.Privilegios != null && !usuario.Privilegios.Contains(((int)Privilegios.Departamentos).ToString()))
             {
                 Response.Redirect("AccessDenied.aspx");
             }
